fix: rename Quote > MessageQuote
